Added Layout and Button components, changed stylesheets to scss
