+++ package.json
@@ -5,6 +5,8 @@
   "dependencies": {
     "@testing-library/jest-dom": "5.16.2",
     "@testing-library/react": "12.1.4",
+    "postcss": "^8.4.12",
+    "postcss-scss": "^4.0.3",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-router-dom": "^6.2.2",
