+++ src/components/layout.scss
@@ -1,0 +1,14 @@
+@import url('https://fonts.googleapis.com/css2?family=Creepster&display=swap');
+
+* {
+  margin: 0;
+  padding: 0;
+  font-family: 'Creepster', cursive;
+}
+.banner {
+  width: 100%;
+  height: 100vh;
+  background-image:linear-gradient(rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5)), url(img/background.jpg);
+  background-size: cover;
+  background-position: center; 
+}
