+++ src/pages/pages.scss
@@ -1,0 +1,11 @@
+.content {
+  width: 100%;
+  position: absolute;
+  top: 50%;
+  transform: translateY(-80%);
+  text-align: center;
+  color: #4AADC9;
+  h1 {
+    font-size: 90px;
+  }
+}
